Set column widths based on font size and style
Uses glyph widths from the Arial and Calibri fonts to set column widths
based on the font size and style of the header row.
This is exact for Google Sheets, but is approximate for Excel since it
assumes the default font is the same as the requested font family
in regular style and size 11.